Flask_app/requirements.txt
alembic==1.3.2
amqp==2.6.1
aniso8601==8.0.0
APScheduler==3.6.3
astroid==2.3.2
atomicwrites==1.4.0
attrs==19.3.0
Babel==2.9.1
bcrypt==3.1.7
billiard==3.6.4.0
bleach==3.3.0
blinker==1.4
Bootstrap-Flask==1.2.0
cached-property==1.5.2
cachetools==4.2.2
celery==4.4.2
certifi==2020.12.5
cffi==1.13.2
chardet==4.0.0
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.1
cryptography==2.8
cycler==0.11.0
dnspython==2.1.0
dominate==2.4.0
email-validator==1.1.2
et-xmlfile==1.0.1
Faker==8.8.1
Flask==1.1.2
Flask-Babel==2.0.0
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-Caching==1.8.0
Flask-Cors==3.0.8
Flask-DebugToolbar==0.10.1
Flask-FontAwesome==0.1.2
Flask-JWT==0.3.2
Flask-Limiter==1.4
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
flask-mongoengine==1.0.0
Flask-PyMongo==2.3.0
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-WkHTMLtoPDF==0.1.0
Flask-WTF==0.14.3
fonttools==4.30.0
fpdf==1.7.2
idna==2.10
importlib-metadata==1.3.0
iniconfig==1.1.1
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.10.3
joblib==1.1.0
jsonschema==3.2.0
kiwisolver==1.3.2
kombu==4.6.11
lazy-object-proxy==1.4.2
limits==1.5.1
linecache2==1.0.0
Mako==1.1.0
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.5.1
mccabe==0.6.1
mongoengine==0.22.1
more-itertools==8.0.2
numpy==1.20.2
openpyxl==3.0.6
packaging==20.9
pandas==1.1.5
pdfkit==1.0.0
Pillow==8.4.0
pluggy==1.0.0
prompt-toolkit==3.0.22
psycopg2==2.8.6
py==1.11.0
pycparser==2.19
PyJWT==1.4.2
pylint==2.4.3
pymongo==3.11.2
pyparsing==2.4.7
PyPDF2==1.26.0
pyrsistent==0.15.6
pytest==7.0.1
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
python-telegram-bot==13.6
pytz==2019.3
reportlab==3.6.2
requests==2.25.1
scikit-learn==1.0.1
scipy==1.7.1
sentry-sdk==0.20.3
shortuuid==1.0.1
six==1.12.0
sklearn==0.0
SQLAlchemy==1.3.10
stripe==2.60.0
telegram==0.0.1
text-unidecode==1.3
threadpoolctl==3.0.0
tomli==2.0.1
tornado==6.1
tqdm==4.62.3
traceback2==1.4.0
typed-ast==1.4.0
tzlocal==2.1
unittest2==1.1.0
urllib3==1.26.3
vine==1.3.0
visitor==0.1.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==0.16.0
wrapt==1.11.2
WTForms==2.3.3
xlrd==1.2.0
xlwt==1.3.0
zipp==0.6.0
